chore: refactor package function
