+++ src/package/index.ts
@@ -1,3 +1,9 @@
-export { applyPolyfill } from './polyfill/CustomElementHMR';
-export { clearDOM } from './helper/ClearDOM';
-export { DefineCustomElement } from './decorator/DefineCustomElement';
+import { initCache } from './polyfill/hmrCache';
+import { overrideCustomElementDefine } from './polyfill/overrideCustomElementDefine';
+export { clearDOM } from './utils/clearDOM';
+export { defineCustomElement } from './decorator/defineCustomElement';
+
+export function applyPolyfill() {
+    initCache();
+    overrideCustomElementDefine();
+}
