+++ src/package/polyfill/createHookClass.ts
@@ -1,0 +1,37 @@
+import { getMostRecentImpl } from './hmrCache';
+
+export function createHookClass(elementName: string, originalImpl: any) {
+    return class extends originalImpl {
+        static get observedAttributes() {
+            return super.observedAttributes;
+        }
+
+        connectedCallback() {
+            const mostRecentImpl = getMostRecentImpl(elementName).prototype;
+            if (mostRecentImpl.connectedCallback) {
+                mostRecentImpl.connectedCallback.apply(this, arguments);
+            }
+        }
+
+        disconnectedCallback() {
+            const mostRecentImpl = getMostRecentImpl(elementName).prototype;
+            if (mostRecentImpl.disconnectedCallback) {
+                mostRecentImpl.disconnectedCallback.apply(this, arguments);
+            }
+        }
+
+        adoptedCallback() {
+            const mostRecentImpl = getMostRecentImpl(elementName).prototype;
+            if (mostRecentImpl.adoptedCallback) {
+                mostRecentImpl.adoptedCallback.apply(this, arguments);
+            }
+        }
+
+        attributeChangedCallback() {
+            const mostRecentImpl = getMostRecentImpl(elementName).prototype;
+            if (mostRecentImpl.attributeChangedCallback) {
+                mostRecentImpl.attributeChangedCallback.apply(this, arguments);
+            }
+        }
+    };
+}
