+++ src/package/utils/clearDOM.ts
